secret set

+++ interview-bank-server/src/question/question.controller.ts
@@ -2,18 +2,16 @@ import { Controller, Get, Post, Body, Query } from '@nestjs/common';
 import { QuestionService } from './question.service';
 import { QuestionDto } from './dto/question.dto';
 import { getRandomID } from '../common/utils';
+import { secretKey } from '../common/common';
 
 @Controller('question')
 export class QuestionController {
   constructor(private questionService: QuestionService) {}
 
-  // @Get('getClass')
-  // getClass(): any {
-  //   return this.questionService.getClass();
-  // }
-
   @Post('saveQuestion')
   async saveQuestion(@Body() questionDto: QuestionDto) {
+    if (questionDto.key !== secretKey)
+      return { state: 'err', errmsg: '身份验证失败' };
     if (!questionDto.code) questionDto.code = getRandomID();
     await this.questionService.save(questionDto);
     return { state: 'ok' };
@@ -21,6 +19,8 @@ export class QuestionController {
 
   @Post('delQuestion')
   async delQuestion(@Body() param: any) {
+    if (param.key !== secretKey)
+      return { state: 'err', errmsg: '身份验证失败' };
     if (!param.code) return { state: 'err', errmsg: '缺少参数code' };
     await this.questionService.delete(param.code);
     return { state: 'ok' };
